Introduce Makefile to consolidate project tooling
The intent of this change is to start using Makefile to build future
project tooling and orchestration which makes easier to work with the
project by making explicit environment variables that often are
implicit and may not be defined in the developers computers.

The makefile will work out of the box for Unix based systems which
include Mac OS, Ubuntu, Fedora and other Linux distributions, since
they all ship with GNU Make (via `make` command).

In this changeset, we set the default target of `make` to
`docker/app/run` which will execute `docker/app/build` target to
build the Docker image needed.

When you want to stop and remove the container, you can run
`make docker/app/stop` and it will stop the Docker container.

Other noteworthy addition is the `DOCKER_BUILDKIT` variable set to
`1` which is assigned in case it is not present in the environment.
This enables a more optimized build process of the Docker image.